pyproject.toml
[project]
name = "ecommerce_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "fastapi>=0.116.1",
    "hf-xet>=1.1.5",
    "ipykernel>=6.30.0",
    "ipython>=9.4.0",
    "langchain-community>=0.3.27",
    "langchain-text-splitters>=0.3.9",
    "langchain[groq]>=0.3.27",
    "langfuse>=3.2.1",
    "langgraph>=0.5.4",
    "numpy>=1.26.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=2.7.0",
    "transformers>=4.51.0",
    "uvicorn>=0.35.0",
]
